updated Spring

+++ src/main/java/com/leeraar/repositories/AuthorityRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * The Interface that executes queries on the database.
  */
-public interface AuthorityRepository extends CrudRepository<Authority, Integer> {
+public interface AuthorityRepository extends CrudRepository<Authority, Long> {
 
 }
